+++ packages/diff-engine/src/index.ts
@@ -13,6 +13,8 @@ import {
   type TextBlock
 } from "@one-shot-ui/core";
 import { detectBackgroundColor, loadImage } from "@one-shot-ui/image-io";
+
+export * from "./spacing.js";
 import { clusterComponents } from "@one-shot-ui/vision-components";
 import { detectLayoutBoxes, measureSpacing } from "@one-shot-ui/vision-layout";
 import { detectGradient, detectShadow, estimateBorderRadius, estimateNodeFill, extractDominantColors } from "@one-shot-ui/vision-style";