+++ packages/vision-layout/src/index.ts
@@ -1,6 +1,9 @@
 import type { LayoutNode, LayoutStrategy, SpacingMeasurement } from "@one-shot-ui/core";
 import { rgbToHex, samplePixel, type ImageAsset } from "@one-shot-ui/image-io";
 
+export * from "./rulers.js";
+export * from "./icons.js";
+
 const GRID_SIZE = 8;
 
 /**